Fix difficulty for solo mining
Must do 64bit compare (or just check for top 32bits being zero,
but a 64bit op is faster).

+++ stratum.h
@@ -6,7 +6,7 @@
 
 typedef struct _JobInfo
 {
-	uint32_t XMRTarget;
+	uint64_t XMRTarget;
 	uint8_t ID[32];
 	uint8_t XMRBlob[128];
 	uint32_t XMRBlobLen;
